tests/wpt/mozilla/tests/mozilla/bluetooth/getPrimaryServices/blacklisted-services.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.blacklist);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [device_information.name]}],
        optionalServices: [
            blacklist_test_service_uuid, generic_access.name,
            heart_rate.name, human_interface_device.name
        ]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryServices())
    .then(services => {
        assert_equals(services.length, 4);
        assert_equals(services[0].uuid, blacklist_test_service_uuid);
        assert_equals(services[1].uuid, device_information.uuid);
        assert_equals(services[2].uuid, generic_access.uuid);
        assert_equals(services[3].uuid, heart_rate.uuid);
    })
    .catch(error => assert_unreached('Promise should have resolved. ' + error));
}, 'Request for services. Does not return blacklisted service.');
</script>
